+++ packages/deprecated/theming-react-native/src/ThemeProvider.types.ts
@@ -1,0 +1,8 @@
+/* eslint-disable */
+
+import { ThemeRegistry } from './Theme.types';
+
+export interface IThemeProviderProps extends React.PropsWithChildren<{}> {
+  theme?: string;
+  registry?: ThemeRegistry;
+}