+++ packages/experimental/Stack/src/Stack.tsx
@@ -7,6 +7,16 @@ import { filterViewProps } from '@fluentui-react-native/adapters';
 import { compose, withSlots, mergeProps, UseSlots, getMemoCache } from '@fluentui-react-native/framework';
 import { stylingSettings } from './Stack.styling';
 
+// Needed for TS to understand that __jsiExecutorDescription exists.
+declare global {
+  /* eslint-disable-next-line @typescript-eslint/no-namespace*/
+  namespace NodeJS {
+    interface Global {
+      __jsiExecutorDescription: any;
+    }
+  }
+}
+
 const mixinCache = getMemoCache<ViewProps>();
 
 /**
@@ -41,10 +51,10 @@ export const Stack = compose<StackType>({
     const { gap, horizontal, wrap, ...rest } = props;
     const Slots = useSlots(props);
     return (final: StackProps, ...children: React.ReactNode[]) => {
-      if (gap && gap > 0 && children) {
+      if (gap && gap > 0 && children && global.__jsiExecutorDescription !== 'ChakraRuntime') {
         const mixinProps = getMixinProps(horizontal, gap);
 
-        /* eslint-disable @typescript-eslint/ban-ts-ignore */
+        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         // @ts-ignore - TODO, fix typing error
         children = React.Children.map(children, (child: React.ReactChild, index: number) => {
           if (React.isValidElement(child) && index > 0) {