+++ packages/experimental/use-slots/src/index.ts
@@ -1,4 +1,0 @@
-export * from './buildUseSlots';
-export * from './stagedComponent';
-export * from './renderSlot';
-export * from './withSlots';