+++ packages/framework/use-slot/src/index.ts
@@ -1,0 +1,4 @@
+export * from './renderSlot';
+export * from './stagedComponent';
+export * from './useSlot';
+export * from './withSlots';