+++ scripts/src/utils/resolvePaths.ts
@@ -1,40 +1,17 @@
-'use strict';
-
-import { findGitRoot, normalizeToUnixPath } from 'just-repo-utils';
+import findUp from 'find-up';
 import fs from 'fs';
 import path from 'path';
 
 export function nodeModulesToRoot(): string[] {
   const results = [];
-  findGitRoot(cur => {
-    const nodeModulePath = path.join(cur, 'node_modules');
-    if (fs.existsSync(nodeModulePath)) {
-      results.push(nodeModulePath.replace(/\\/g, '/'));
+  findUp.sync((directory) => {
+    const nodeModulesPath = path.join(directory, 'node_modules');
+    if (fs.existsSync(nodeModulesPath)) {
+      results.push(nodeModulesPath.replace(/\\/g, '/'));
     }
-    return false;
+
+    const gitRoot = path.join(directory, '.git');
+    return fs.existsSync(gitRoot) ? findUp.stop : undefined;
   });
   return results;
-}
-
-function queryModule(name: string, direct?: boolean): string {
-  const cur = direct
-    ? path.resolve(require.resolve(name, { paths: [process.cwd()] }))
-    : path.resolve(require.resolve(name + '/package.json', { paths: [process.cwd()] }), '..');
-  return normalizeToUnixPath(fs.realpathSync(cur));
-}
-
-/**
- * Resolve a module to a true, normalized, non-symlink path
- * @param moduleName - name of the module to resolve
- */
-export function resolveModule(moduleName: string): string {
-  return queryModule(moduleName);
-}
-
-/**
- * Resolve a file reference to a true, normalized, non-symlink path
- * @param fileName - file to resolve
- */
-export function resolveFile(fileName: string) {
-  return queryModule(fileName, true);
 }